Moved everything added to silx.gui.plot into silx.gui.data._RecordPlot
